requirements.txt
pywikibot
pyyaml
